refactor(fe): 유사도 로직 파일명 및 함수명 리팩토링
