+++ client/src/features/similarity/index.ts
@@ -1,7 +1,3 @@
 export type { Similarity } from './model/similarity';
 export type { PreprocessedStrokeData } from './model/preprocessedStrokeData';
-export {
-  preprocessStrokes,
-  calculateFinalSimilarityByStrokes,
-  calculateFinalSimilarityByPreprocessed,
-} from './lib/calculateFinalSimilarity';
+export { preprocessStrokes, scoreFinalSimilarity } from './lib/final';
